refactor: Move IKButton to MailCoreUI
