+++ Mail/Components/Custom Buttons/ExtendedFAB.swift
@@ -16,6 +16,7 @@
  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
+import MailCoreUI
 import MailResources
 import SwiftUI
